added: RegisterTravelNewsEditerView
